source_code/back-end/requirements.txt
flask~=2.3.3
mysql-connector-python~=8.2.0
python-dateutil~=2.8.2
pytz~=2022.4